# vigia-services/apps/vigia-api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API do Vigia Services.",
        "title": "Vigia API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
        "/health": {
            "get": {
                "description": "Verifica se a API está em execução.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_http_handlers.HealthCheckResponse"
                        }
                    }
                }
            }
        },
        "/v1/devices/": {
            "get": {
                "description": "Retorna a lista de dispositivos cadastrados para o usuário autenticado.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "Listar dispositivos do usuário",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_http_handlers.MessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_http_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/devices/register": {
            "post": {
                "description": "Vincula um dispositivo a um usuário autenticado.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "Registrar dispositivo do usuário",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_http_handlers.MessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_http_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/devices/version/find-for-updates": {
            "get": {
                "description": "Retorna metadados da versão mais recente publicada. Quando `currentVersion` é informada via query, responde 204 caso o dispositivo já esteja na versão mais recente.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "versions"
                ],
                "summary": "Buscar atualização disponível",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Versão atualmente instalada no dispositivo (semver)",
                        "name": "currentVersion",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_PedroReis16_vigia_vigia-services_apps_vigia-api_internal_models_dtos.VersionDTO"
                        }
                    },
                    "204": {
                        "description": "Dispositivo já está na versão mais recente"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_http_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/devices/version/register": {
            "post": {
                "description": "Faz upload de um bundle do agente fall-detection e registra uma nova versão no catálogo.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "versions"
                ],
                "summary": "Registrar nova versão do Vigia",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Identificador semântico da versão (ex.: 1.2.3)",
                        "name": "version",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Bundle compactado da versão",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_http_handlers.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_http_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_http_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/devices/version/{version}/download": {
            "get": {
                "description": "Retorna a URL de download pré-assinada para a versão informada.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "versions"
                ],
                "summary": "Obter metadados de uma versão",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Versão alvo (semver)",
                        "name": "version",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_PedroReis16_vigia_vigia-services_apps_vigia-api_internal_models_dtos.VersionDTO"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_http_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_http_handlers.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_PedroReis16_vigia_vigia-services_apps_vigia-api_internal_models_dtos.VersionDTO": {
            "type": "object",
            "properties": {
                "download_url": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "internal_http_handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "internal_http_handlers.HealthCheckResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "internal_http_handlers.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    }
}